+++ src/type/fantasy-land.js
@@ -1,42 +1,55 @@
 //@flow
 
 import { equals, merge } from 'ramda'
-import { copyProps } from '../decorators'
+import { append, cons } from '../utils/list'
+import { type TypeRecord, type TypeStatic } from './type-container'
 
 
 export const fantasyMethods = {
-  equals: (ctx, Record) => (val): boolean => {
+  equals: (ctx: TypeRecord<any>, Record: TypeStatic<any>) => (val: TypeRecord<any>): boolean => {
     if (!Record.is(val)) return false
     const asTyped = Record.of(val)
     const json = ctx.toJSON()
     return equals(json, asTyped.toJSON())
   },
-  concat: (ctx, Record) => (val) => {
+  concat: (ctx: TypeRecord<any>, Record: TypeStatic<any>) => (val: TypeRecord<any>) => {
     const struct = Record
       .of(val)
       .toJSON()
     return Record(merge(ctx.toJSON(), struct))
   },
-  extract: (ctx) => function extract() { return ctx.toJSON() },
-  ap     : (ctx) => function ap(m) { return m.chain(f => ctx.map(f)) },
-  extend : (ctx, Record) => function extend(fn) { return Record(fn(ctx)) },
-  chain  : (ctx) => (fn) => fn(ctx.toJSON()),
-  map    : (ctx, Record) => (fn) => Record(fn(ctx)),
+  extract: (ctx: TypeRecord<any>) => function extract() { return ctx.toJSON() },
+  ap     : (ctx: TypeRecord<any>) =>
+    function ap(m: TypeRecord<any>) { return m.chain(f => ctx.map(f)) },
+  extend : (ctx: TypeRecord<any>, Record: TypeStatic<any>) =>
+    function extend(fn: (val: TypeRecord<any>) => any) { return Record(fn(ctx)) },
+  chain  : (ctx: TypeRecord<any>) =>
+    (fn: (val: any) => TypeRecord<any>) =>
+      fn(ctx.toJSON()),
+  map    : (ctx: TypeRecord<any>, Record: TypeStatic<any>) =>
+    (fn: (val: any) => TypeRecord<any>) =>
+      Record(fn(ctx)),
 }
 
 export const fantasyStatic = {
-  of: (Record) => (val: *) =>
+  of: (Record: TypeStatic<any>) => (val: *) =>
     val && (Record.ಠ_ಠ === val.ಠ_ಠ)
       ? val
       : Record(val),
-  equals   : () => (a, b) => a.equals(b),
-  contramap: (Record) => (fn) => {
-    function preprocessed(...data: S[]) {
-      const preprocessResult = fn(...data)
-      return Record(preprocessResult)
-    }
-    copyProps(Record, preprocessed) //TODO replace copyProps
-    return preprocessed
+  equals   : () => (a: TypeRecord<any>, b: TypeRecord<any>) => a.equals(b),
+  contramap: (Record: TypeStatic<any>) => (fn: *) => {
+    //$FlowIssue
+    const newStack = append(fn, Record.stack)
+    //$FlowIssue
+    const NewRecord = Record.stackUpdate(newStack)
+    return NewRecord
+  },
+  map: (Record: TypeStatic<any>) => (fn: *) => {
+    //$FlowIssue
+    const newStack = cons(fn, Record.stack)
+    //$FlowIssue
+    const NewRecord = Record.stackUpdate(newStack)
+    return NewRecord
   },
 }