+++ src/validation/index.js
@@ -1,6 +1,0 @@
-//@flow
-
-import Check from './check'
-import Rule from './rule'
-
-export { Check, Rule }